+++ packages/orq-rc/src/models/operations/createagentresponserequest.ts
@@ -587,8 +587,9 @@ export const CreateAgentResponseRequestResponse$inboundSchema: z.ZodType<
   z.instanceof(ReadableStream<Uint8Array>)
     .transform(stream => {
       return new EventStream(stream, rawEvent => {
-        if (rawEvent.data === "[DONE]") return { done: true };
+        if (rawEvent.data === "[DONE]") return { done: true, value: undefined };
         return {
+          done: false,
           value: z.lazy(() =>
             CreateAgentResponseRequestResponseBody$inboundSchema
           ).parse(rawEvent),