+++ packages/orq-rc/src/models/operations/createresponse.ts
@@ -3303,8 +3303,9 @@ export const CreateResponseResponse$inboundSchema: z.ZodType<
   z.instanceof(ReadableStream<Uint8Array>)
     .transform(stream => {
       return new EventStream(stream, rawEvent => {
-        if (rawEvent.data === "[DONE]") return { done: true };
+        if (rawEvent.data === "[DONE]") return { done: true, value: undefined };
         return {
+          done: false,
           value: z.lazy(() =>
             CreateResponseRouterResponsesResponseBody$inboundSchema
           ).parse(rawEvent),